+++ lib/src/animation_panel.dart
@@ -7,10 +7,8 @@ class AnimationPanel<T> extends StatelessWidget {
     required this.width,
     required this.height,
     required this.hasTouch,
-    this.elevation,
     this.duration,
     this.radius = 24,
-    this.color,
     this.curve,
     this.decoration,
   }) : super(key: key);
@@ -19,9 +17,7 @@ class AnimationPanel<T> extends StatelessWidget {
   final double? width;
   final double? height;
   final double radius;
-  final double? elevation;
   final Duration? duration;
-  final Color? color;
   final Curve? curve;
   final bool hasTouch;
   final BoxDecoration? decoration;
@@ -34,15 +30,21 @@ class AnimationPanel<T> extends StatelessWidget {
     return AnimatedContainer(
       transform: Matrix4.translationValues(_offset, 0, 0),
       duration: hasTouch == false
-          ? const Duration()
+          ? Duration.zero
           : duration ?? const Duration(milliseconds: 200),
       curve: curve!,
       width: width,
       decoration: decoration?.copyWith(
-        borderRadius: decoration?.borderRadius ?? BorderRadius.circular(radius),
-        color: decoration?.color ?? color,
-      ),
-      child: Container(height: height),
+            borderRadius:
+                decoration?.borderRadius ?? BorderRadius.circular(radius),
+            color: decoration?.color ?? Colors.white,
+          ) ??
+          BoxDecoration(
+            borderRadius:
+                decoration?.borderRadius ?? BorderRadius.circular(radius),
+            color: decoration?.color ?? Colors.white,
+          ),
+      height: height,
     );
   }
 }